# application/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mechanic Shop API",
        "description": "RESTful API for managing customers, mechanics, service tickets, and inventory.",
        "version": "3.0.0"
    },
    "host": "127.0.0.1:5001",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter: Bearer followed by the JWT token"
        }
    },
...
    "paths": {
...
        "/mechanics/": {
            "post": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Create a mechanic",
                "description": "Creates a new mechanic after validating the request body and confirming that the email is not already associated with another mechanic.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Mechanic information",
                        "schema": {
                            "$ref": "#/definitions/MechanicPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mechanic created successfully",
                        "schema": {
                            "$ref": "#/definitions/MechanicResponse"
                        },
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "name": "Arthur Fonzarelli",
                                "email": "thefonz@mechanics.com",
                                "phone": "5559876543",
                                "salary": "75000"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or duplicate email",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "error": "Email already associated with a mechanic."
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Get all mechanics",
                "description": "Returns a list of all mechanics stored in the database.",
                "responses": {
                    "200": {
                        "description": "Mechanics retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/MechanicResponse"
                            }
                        },
                        "examples": {
                            "application/json" : [
                                {
                                    "id": 1,
                                    "name": "Arthur Fonzarelli",
                                    "email": "thefonz@mechanics.com",
                                    "phone": "5559876543",
                                    "salary": "75000"
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mechanics/login": {
            "post": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Mechanic login",
                "description": "Validates mechanic credentials and returns a mechanic-specific JWT Bearer Token.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Mechanic login credentials",
                        "schema": {
                            "$ref": "#/definitions/LoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mechanic login successful",
                        "schema": {
                            "$ref": "#/definitions/MechanicLoginResponse"
                        },
                        "examples": {
                            "application/json": {
                                "message": "Mechanic login successful",
                                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "error": "Invalid email or password"
                            }
                        }
                    }
                }
            }
        },
...
        "/mechanics/{mechanic_id}": {
            "parameters": [
                {
                    "name": "mechanic_id",
                    "in": "path",
                    "required": true,
                    "type": "integer",
                    "description": "Mechanic ID"
                }
            ],
            "get": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Get mechanic by ID",
                "description": "Retrieves a single mechanic using the ID provided in the URL.",
                "responses": {
                    "200": {
                        "description": "Mechanic retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/MechanicResponse"
                        }
                    },
                    "404": {
                        "description": "Mechanic not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "error": "Mechanic not found"
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Update mechanic",
                "description": "Updates the authenticated mechanic's information. The mechanic ID stored in the token must match the mechanic ID in the URL.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Mechanic fields to update",
                        "schema": {
                            "$ref": "#/definitions/MechanicUpdatePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mechanic updated successfully",
                        "schema": {
                            "$ref": "#/definitions/MechanicResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or duplicate email",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid mechanic token",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "error": "Token is missing"
                            }
                        }
                    },
                    "403": {
                        "description": "Authenticated mechanic does not match the requested mechanic",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        },
                        "examples": {
                            "application/json": {
                                "error": "Unauthorized"
                            }
                        }
                    },
                    "404": {
                        "description": "Mechanic not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mechanics"
                ],
                "summary": "Delete mechanic",
                "description": "Deletes the authenticated mechanic and removes their service-ticket relationships without deleting the service tickets.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mechanic deleted successfully",
                        "examples": {
                            "application/json": {
                                "message": "Mechanic deleted successfully"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid mechanic token",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Mechanic not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "LoginPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "rachel@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Customer not found"
                }
            }
        },
...
        "MechanicPayload": {
            "type": "object",
            "required": [
                "name",
                "email",
                "password",
                "phone",
                "salary"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Arthur Fonzarelli"
                },
                "email": {
                    "type": "string",
                    "example": "thefonz@mechanics.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "phone": {
                    "type": "string",
                    "example": "5559876543"
                },
                "salary": {
                    "type": "number",
                    "format": "float",
                    "example": 75000
                }
            }
        },
        "MechanicUpdatePayload": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Arthur Fonzarelli"
                },
                "email": {
                    "type": "string",
                    "example": "updated@mechanics.com"
                },
                "password": {
                    "type": "string",
                    "example": "newpassword123"
                },
                "phone": {
                    "type": "string",
                    "example": "5551112222"
                },
                "salary": {
                    "type": "number",
                    "format": "float",
                    "example": 82000
                }
            }
        },
        "MechanicResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Arthur Fonzarelli"
                },
                "email": {
                    "type": "string",
                    "example": "thefonz@mechanics.com"
                },
                "phone": {
                    "type": "string",
                    "example": "5559876543"
                },
                "salary": {
                    "type": "number",
                    "format": "float",
                    "example": 75000
                }
            }
        },
        "MechanicLoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Mechanic login successful"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
    }
}